avs2/avs3: CB2 Type mismatch causes compilation to fail. [1/1]
PD#SWPL-211933

Problem:
Type mismatch causes compilation to fail.

Solution:
Cast numeric literals(16) to u32.

Verify:
ohm

+++ drivers/frame_provider/decoder/avs2_fb/vavs2_fb.c
@@ -4880,7 +4880,7 @@ static u32 get_dynamic_buf_num_margin(struct AVS2Decoder_s *dec)
 		(dynamic_buf_num_margin & 0x7fffffff);
 }
 
-static u32 avs2_calc_dbp_size_with_level_id(int level_id, int MinCuWidth, int MinCuHeight) {
+static u32 avs2_calc_dbp_size_with_level_id(uint16_t level_id, u32 MinCuWidth, u32 MinCuHeight) {
 	u32 dbp_size = 15;
 	u32 frame_size = MinCuWidth * AVS2_MINI_SIZE * MinCuHeight * AVS2_MINI_SIZE;
 	switch (level_id)
@@ -4898,7 +4898,7 @@ static u32 avs2_calc_dbp_size_with_level_id(int level_id, int MinCuWidth, int Mi
 	case AVS2_LEVEL_ID_6_2_60:
 	case AVS2_LEVEL_ID_6_0_120:
 	case AVS2_LEVEL_ID_6_2_120:
-		dbp_size = min((13369344 + frame_size - 1) / frame_size, 16) - 1;
+		dbp_size = min((13369344 + frame_size - 1) / frame_size, (u32)16) - 1;
 		break;
 	case AVS2_LEVEL_ID_8_0_30:
 	case AVS2_LEVEL_ID_8_2_30:
@@ -4906,7 +4906,7 @@ static u32 avs2_calc_dbp_size_with_level_id(int level_id, int MinCuWidth, int Mi
 	case AVS2_LEVEL_ID_8_2_60:
 	case AVS2_LEVEL_ID_8_0_120:
 	case AVS2_LEVEL_ID_8_2_120:
-		dbp_size = min((56623104 + frame_size - 1) / frame_size, 16) - 1;
+		dbp_size = min((56623104 + frame_size - 1) / frame_size, (u32)16) - 1;
 		break;
 	case AVS2_LEVEL_ID_10_0_30:
 	case AVS2_LEVEL_ID_10_2_30:
@@ -4914,7 +4914,7 @@ static u32 avs2_calc_dbp_size_with_level_id(int level_id, int MinCuWidth, int Mi
 	case AVS2_LEVEL_ID_10_2_60:
 	case AVS2_LEVEL_ID_10_0_120:
 	case AVS2_LEVEL_ID_10_2_120:
-		dbp_size = min((213909504  + frame_size - 1) / frame_size, 16) - 1;
+		dbp_size = min((213909504  + frame_size - 1) / frame_size, (u32)16) - 1;
 		break;
 	default:
 		break;
@@ -4923,8 +4923,8 @@ static u32 avs2_calc_dbp_size_with_level_id(int level_id, int MinCuWidth, int Mi
 }
 
 static u32 avs2_calc_dpb_size(struct AVS2Decoder_s *dec) {
-	int MinCuHeight = 0;
-	int MinCuWidth = 0;
+	u32 MinCuHeight = 0;
+	u32 MinCuWidth = 0;
 
 	if (max_buf_num > 0) {
 		avs2_print(dec, 0, "[%s]  use custom dbp size:%d\n", __func__, max_buf_num);
